+++ editxt/test/test_commands.py
@@ -461,6 +461,7 @@ def test_set_variable():
         class document:
             class syntaxdef:
                 name = "Unknown"
+                comment_token = "x"
         text_view = object
         dirname = lambda:None
         font = Font("Mension", 15.0, False, None)
@@ -473,6 +474,7 @@ def test_set_variable():
         eq_(bar.get_completions(command), (completions, None))
         eq_(bar.get_placeholder(command), placeholder)
     yield test, "set ", [
+            "comment_token",
             "font",
             "highlight_selected_text",
             "indent",
@@ -563,9 +565,21 @@ def test_set_variable():
             editor.text_view = TestConfig(textContainer=lambda:None)
             do = CommandTester(mod.set_variable, editor=editor)
             do(command)
-            eq_(getattr(editor, "syntaxdef"), lang)
+            eq_(editor.syntaxdef, lang)
     yield test, "set language py", "Python"
     yield test, "set lang plain", "Plain Text"
+
+    # set comment_token
+    def test(command, token):
+        with test_app("editor*") as app:
+            editor = app.windows[0].current_editor
+            editor.text_view = TestConfig(textContainer=lambda:None)
+            do = CommandTester(mod.set_variable, editor=editor)
+            do(command)
+            eq_(editor.document.comment_token, token)
+    yield test, "set comment_token #", "#"
+    yield test, "set comment_token //", "//"
+    yield test, "set comment_token", "//"
 
 def test_panel_actions():
     from editxt.editor import Editor